% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossJoin}
\alias{crossJoin}
\title{Cross join}
\usage{
crossJoin(x, y)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A data frame.}
}
\value{
A \code{\link[data.table]{data.table}} containing the Cartesian
product of the rows from \code{x} and \code{y}.
}
\description{
Uses \code{\link[data.table]{data.table}} to efficiently perform a cross
join between two data frames; that is, the Cartesian product of the rows
from two data frames.
}
\keyword{internal}
