% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data}
\alias{get_training_data}
\title{#' @keywords internal
where <- function(name, env = parent.frame()) {
  if (identical(env, emptyenv())) {
    stop("Can't find ", name, ".", call. = FALSE)
  } else if (exists(name, envir = env, inherits = FALSE)) {
    env
  } else {
    where(name, parent.env(env))
  }
}}
\usage{
get_training_data(object)
}
\description{
#' @keywords internal
where <- function(name, env = parent.frame()) {
  if (identical(env, emptyenv())) {
    stop("Can't find ", name, ".", call. = FALSE)
  } else if (exists(name, envir = env, inherits = FALSE)) {
    env
  } else {
    where(name, parent.env(env))
  }
}
}
\keyword{internal}
