% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_2d.R
\name{partial_2d}
\alias{partial_2d}
\alias{partial_2d.default}
\title{Partial Dependence (Two Variables)}
\usage{
partial_2d(object, ...)

\method{partial_2d}{default}(object, x1.name, x2.name, n1, n2, x1.values,
  x2.values, super.type, which.class = 1L, check.class = TRUE, newdata, ...)
}
\arguments{
\item{object}{A \code{mertree} object.}

\item{...}{Additional optional arguments to be passed onto \code{aaply}.}

\item{x1.name}{Character string giving the name of the first independent
variable of interest.}

\item{x2.name}{Character string giving the name of the second independent
variable of interest.}

\item{n1}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x1.name}.}

\item{n2}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x2.name}.}

\item{x1.values}{Vector of predictor values to use in computing the partial 
dependence function.}

\item{x2.values}{Vector of predictor values to use in computing the partial 
dependence function.}

\item{super.type}{Character string specifying the type of supervised
learning. Current options are \code{"regression"} or 
\code{"classification"}. For tree-based models (e.g., \code{"rpart"}), the
function can usually extract the necessary information from \code{object}.}

\item{which.class}{Integer specifying which class to use as the target class
(classification only).}

\item{check.class}{Logical indicating whether or not to check the class of
each predictor variable of interest.}

\item{newdata}{An optional data frame.}
}
\description{
One- and two-dimensional partial dependence plots for objects of class
\code{mertree}.
}

