% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartial.R
\name{plotPartial}
\alias{plotPartial}
\alias{plotPartial.partial}
\title{Plotting Partial Dependence Functions}
\usage{
plotPartial(x, ...)

\method{plotPartial}{partial}(x, smooth = FALSE, rug = FALSE,
  chull = FALSE, levelplot = TRUE, contour = FALSE, number = 4,
  overlap = 0.1, train = NULL, col.regions = viridis::viridis, ...)
}
\arguments{
\item{x}{An object that inherits from the \code{"partial"} class.}

\item{...}{Additional optional arguments to be passed onto \code{levelplot},
\code{wireframe}, or \code{xyplot}.}

\item{smooth}{Logical indicating whether or not to overlay a LOESS smoother.
Default is \code{FALSE}.}

\item{rug}{Logical indicating whether or not to include a rug representation
to the plot. If \code{TRUE} the user must supply the original training data
via the \code{train} option. Default is \code{FALSE}.}

\item{chull}{Logical indicating whether or not to draw the convex hull
around the first two variables. If \code{TRUE} the user must supply the
original training data via the \code{train} option.Default is \code{FALSE}.}

\item{levelplot}{Logical indicating whether or not to use a false color level
plot (\code{TRUE}) or a 3-D surface (\code{FALSE}). Default is \code{TRUE}.}

\item{contour}{Logical indicating whether or not to add contour lines to the
level plot. Only used when \code{levelplot = TRUE}. Default is
\code{FALSE}.}

\item{number}{Integer specifying the number of conditional intervals for the
panel variables.}

\item{overlap}{The fraction of overlap of the conditioning variables. See
\code{?graphics::co.intervals} and \code{?lattice::equal.count} for further
details.}

\item{train}{Data frame containing the original training data. Only
required if \code{rug = TRUE} or \code{chull = TRUE}.}

\item{col.regions}{Color vector to be used if \code{contour} is \code{TRUE}.
Defaults to the wonderful Matplotlib 'viridis' color map provided by the
\code{viridis} package. See \code{?viridis::viridis} for details.}
}
\description{
Plots partial dependence functions using \code{lattice} graphics.
}

