% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_progress}
\alias{progress_progress}
\title{Progress bar}
\usage{
progress_progress(...)
}
\arguments{
\item{...}{Additional optional arguments to be passed onto
\code{\link[progress]{progress_bar}}.}
}
\description{
More informative progress bar powered by
\code{\link[progress]{progress_bar}}.
}
