% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trellis.last.object}
\alias{trellis.last.object}
\title{Retrieve the last Trellis Object}
\usage{
trellis.last.object(..., prefix)
}
\description{
See \code{\link[lattice]{trellis.last.object}} for more details.
}
\keyword{internal}
