% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial}
\alias{partial}
\alias{partial.default}
\title{Partial Dependence Functions}
\usage{
partial(object, ...)

\method{partial}{default}(object, pred.var, pred.grid, grid.resolution = NULL,
  super.type, which.class = 1L, check.class = TRUE, newdata, plot = FALSE,
  ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional optional arguments to be passed onto \code{aaply}.}

\item{pred.var}{Character string giving the names of the independent
variables of interest.}

\item{super.type}{Character string specifying the type of supervised
learning. Current options are \code{"regression"} or
\code{"classification"}. For tree-based models (e.g., \code{"rpart"}), the
function can usually extract the necessary information from \code{object}.}

\item{which.class}{Integer specifying which column of the matrix of predicted
probabilities to use as the "focus" class. Default is to use the first class.}

\item{check.class}{Logical indicating whether or not to check the class of
the predictor variable of interest. Default is \code{TRUE}.}

\item{newdata}{An optional data frame.}

\item{plot}{Logical indicating whether to return a data frame containing the
partial dependence values (\code{FALSE}) or plot the partial dependence
function directly (\code{TRUE}). Default is \code{FALSE}.}

\item{n.pts}{Integer giving the number of unique data points to use in
computing the partial dependence values.}
}
\description{
Compute partial dependence functions for various model fitting objects.
}

