% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.partial}
\alias{plot.partial}
\alias{plot.ice}
\alias{plot.cice}
\title{Plotting Partial Dependence Functions}
\usage{
\method{plot}{partial}(
  x,
  center = FALSE,
  plot.pdp = TRUE,
  pdp.col = "red2",
  pdp.lwd = 2,
  pdp.lty = 1,
  smooth = FALSE,
  rug = FALSE,
  contour = FALSE,
  contour.color = "white",
  train = NULL,
  alpha = 1,
  legend.title = "yhat",
  ...
)

\method{plot}{ice}(
  x,
  center = FALSE,
  plot.pdp = TRUE,
  pdp.col = "red2",
  pdp.lwd = 2,
  pdp.lty = 1,
  rug = FALSE,
  train = NULL,
  alpha = 1,
  ...
)

\method{plot}{cice}(
  x,
  plot.pdp = TRUE,
  pdp.col = "red2",
  pdp.lwd = 2,
  pdp.lty = 1,
  rug = FALSE,
  train = NULL,
  alpha = 1,
  ...
)
}
\arguments{
\item{x}{An object that inherits from class \code{"partial"}, \code{"ice"},
or \code{"cice"}; typically the result of a call to \code{\link{partial}}.}

\item{center}{Logical indicating whether or not to produce centered ICE
curves (c-ICE curves). Only useful when \code{x} represents a set of ICE
curves; see \code{\link[pdp]{partial}} for details. Default is \code{FALSE}.}

\item{plot.pdp}{Logical indicating whether or not to plot the partial
dependence function on top of the ICE curves. Default is \code{TRUE}.}

\item{pdp.col}{Character string specifying the color to use for the partial
dependence function when \code{plot.pdp = TRUE}. Default is \code{"red2"}.}

\item{pdp.lwd}{Integer specifying the line width to use for the partial
dependence function when \code{plot.pdp = TRUE}. Default is \code{2}. See
\code{\link[graphics]{par}} for more details.}

\item{pdp.lty}{Integer or character string specifying the line type to use
for the partial dependence function when \code{plot.pdp = TRUE}. Default is
\code{1}. See \code{\link[graphics]{par}} for more details.}

\item{smooth}{Logical indicating whether or not to overlay a LOESS smooth.
Default is \code{FALSE}.}

\item{rug}{Logical indicating whether or not to include rug marks (i.e.,
the min/max and deciles of the predictor distribution) on the predictor
axes. Not currently supported for faceted displays (i.e., partial dependence
of two predictors where at least one is a factor). Default is \code{FALSE}.}

\item{contour}{Logical indicating whether or not to add contour lines to the
false color level plot used for two continuous predictors. Default is
\code{FALSE}.}

\item{contour.color}{Character string specifying the color to use for the
contour lines when \code{contour = TRUE}. Default is \code{"white"}.}

\item{train}{Data frame containing the original training data. Only required
if \code{rug = TRUE}.}

\item{alpha}{Numeric value in \code{[0, 1]} specifying the opacity alpha;
most useful when plotting ICE/c-ICE curves. Default is \code{1} (i.e., no
transparency).}

\item{legend.title}{Character string specifying the text for the legend
title of the false color level plot used for two continuous predictors.
Default is \code{"yhat"}.}

\item{...}{Additional optional arguments to be passed on to
\code{\link[tinyplot]{tinyplot}} (e.g., \code{palette}, \code{main}, or
\code{theme}).}
}
\value{
Draws a plot as a side effect and (invisibly) returns \code{x}.
}
\description{
Plot partial dependence functions (i.e., marginal effects) and individual
conditional expectation (ICE) curves using lightweight base R graphics via
the \href{https://grantmcdermott.com/tinyplot/}{tinyplot} package.
}
\examples{
\dontrun{
#
# Regression example (requires randomForest package to run)
#

# Fit a random forest to the Boston housing data
library(randomForest)
data (boston)  # load the boston housing data
set.seed(101)  # for reproducibility
boston.rf <- randomForest(cmedv ~ ., data = boston)

# Partial dependence of cmedv on lstat
pd <- partial(boston.rf, pred.var = "lstat")
plot(pd, rug = TRUE, train = boston)

# Partial dependence of cmedv on lstat and rm
pd2 <- partial(boston.rf, pred.var = c("lstat", "rm"), chull = TRUE)
plot(pd2, contour = TRUE)

# ICE and c-ICE curves
rm.ice <- partial(boston.rf, pred.var = "rm", ice = TRUE)
plot(rm.ice, rug = TRUE, train = boston, alpha = 0.2)
plot(rm.ice, center = TRUE, alpha = 0.2)
}
}
