% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartial.R
\name{plotPartial}
\alias{plotPartial}
\alias{plotPartial.partial_1d}
\alias{plotPartial.partial_2d}
\title{Plotting Partial Dependence Functions}
\usage{
plotPartial(x, ...)

\method{plotPartial}{partial_1d}(x, ...)

\method{plotPartial}{partial_2d}(x, contour = TRUE, ...)
}
\arguments{
\item{x}{An object of class{"partial_1d"} or \code{"partial_2d"}.}

\item{...}{Additional optional arguments to be passed onto \code{lattice}
plotting functions.}

\item{contour}{Logical indicating whether or not to use \code{lattice::levelplot} 
(\code{TRUE}) or \code{lattice::wireframe} (\code{FALSE}). Default is \code{TRUE}.}
}
\description{
Plots 1- and 2-dimensional partial dependence functions.
}

