% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartial.R
\name{plotPartial}
\alias{plotPartial}
\alias{plotPartial.partial}
\title{Plotting Partial Dependence Functions}
\usage{
plotPartial(x, ...)

\method{plotPartial}{partial}(x, smooth = FALSE, contour = TRUE,
  rug = FALSE, convex.hull = FALSE, number = 4, overlap = 0.1,
  training.data = NULL, ...)
}
\arguments{
\item{x}{An object of class{"partial_1d"} or \code{"partial_2d"}.}

\item{...}{Additional optional arguments to be passed onto \code{levelplot},
\code{wireframe}, or \code{xyplot}.}

\item{smooth}{Logical}

\item{contour}{Logical indicating whether or not to use
\code{lattice::levelplot} (\code{TRUE}) or \code{lattice::wireframe}
(\code{FALSE}). Default is \code{TRUE}.}

\item{rug}{Logical indicating whether or not to include a rug representation
to the plot. If \code{TRUE} the user must supply the original data.}

\item{convex.hull}{Logical}

\item{number}{Integer}

\item{overlap}{Proportion}
}
\description{
Plots partial dependence functions using \code{lattice} graphics.
}

